backend/requirements.txt
FastAPI
uvicorn
sqlalchemy
asyncpg
pydantic
python-jose
passlib[bcrypt]
alembic
httpx
python-dotenv